feat: Lv2 유저 CRUD 및 일정-유저 연관관계 구현

+++ src/main/java/com/example/scheduledevelop/schedule/service/ScheduleService.java
@@ -3,6 +3,8 @@ package com.example.scheduledevelop.schedule.service;
 import com.example.scheduledevelop.schedule.dto.*;
 import com.example.scheduledevelop.schedule.entity.Schedule;
 import com.example.scheduledevelop.schedule.repository.ScheduleRepository;
+import com.example.scheduledevelop.user.entity.User;
+import com.example.scheduledevelop.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,12 +16,17 @@ import java.util.List;
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository; //레파지토리 가져옴
+    private final UserRepository userRepository;
 
     // 저장
     @Transactional
     public ScheduleCreateResponse save(ScheduleCreateRequest request) {
+        User savedUser = userRepository.findById(request.getUserid()).orElseThrow(
+                () -> new IllegalStateException("존재하지 않는 유저입니다.")
+        );
+
         Schedule schedule = new Schedule(
-                request.getUsername(),
+                savedUser,
                 request.getTitle(),
                 request.getContent()
         );
@@ -50,18 +57,27 @@ public class ScheduleService {
     // 수정
     @Transactional
     public ScheduleUpdateResponse update(Long scheduleId, ScheduleUpdateRequest request) {
+
         Schedule updatedSchedule = scheduleRepository.findById(scheduleId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 일정입니다.")
         );
+
+        if (!updatedSchedule.getUser().getId().equals(request.getUserid())) {
+            throw new IllegalStateException("작성한 유저만 수정할 수 있습니다.");
+        }
+        updatedSchedule.update(request);
         return ScheduleUpdateResponse.from(updatedSchedule);
     }
 
     // 삭제
     @Transactional
-    public void delete(Long scheduleId) {
+    public void delete(Long scheduleId,ScheduleDeleteRequest request) {
         Schedule deletedSchedule = scheduleRepository.findById(scheduleId).orElseThrow(
                 () -> new IllegalStateException("존재하지 않는 일정입니다.")
         );
+        if (!deletedSchedule.getUser().getId().equals(request.getUserid())) {
+            throw new IllegalStateException("작성한 유저만 삭제할 수 있습니다.");
+        }
         scheduleRepository.delete(deletedSchedule);
     }
